+++ lib/ffi/gphoto2_port/gp_port_serial_parity.rb
@@ -1,0 +1,6 @@
+module FFI
+  module GPhoto2Port
+    # libgphoto2_port/libgphoto2_port/gphoto2-port.h
+    GPPortSerialParity = enum :off, :even, :odd
+  end
+end