+++ lib/ffi/gphoto2_port/gp_port_settings.rb
@@ -1,0 +1,11 @@
+module FFI
+  module GPhoto2Port
+    class GPPortSettings < FFI::Struct
+      # libgphoto2_port/libgphoto2_port/gphoto2-port.h
+      layout :serial, GPPortSettingsSerial,
+             :usb, GPPortSettingsUSB,
+             :usbdiskdirect, GPPortSettingsUsbDiskDirect,
+             :usbscsi, GPPortSettingsUsbScsi
+    end
+  end
+end