+++ lib/ffi/gphoto2_port/gp_port_settings_serial.rb
@@ -1,0 +1,12 @@
+module FFI
+  module GPhoto2Port
+    class GPPortSettingsSerial < FFI::Struct
+      # libgphoto2_port/libgphoto2_port/gphoto2-port.h
+      layout :port, [:char, 128],
+             :speed, :int,
+             :bits, :int,
+             :parity, GPPortSerialParity,
+             :stopbits, :int
+    end
+  end
+end