+++ lib/ffi/gphoto2_port/gp_port_settings_usb_scsi.rb
@@ -1,0 +1,8 @@
+module FFI
+  module GPhoto2Port
+    class GPPortSettingsUsbScsi < FFI::Struct
+      # libgphoto2_port/libgphoto2_port/gphoto2-port.h
+      layout :path, [:char, 128]
+    end
+  end
+end